Move processPlugins to separate module
