+++ src/lib/substituteTailwindAtRules.js
@@ -6,51 +6,8 @@ import utilityModules from '../utilityModules'
 import applyClassPrefix from '../util/applyClassPrefix'
 import escapeClassName from '../util/escapeClassName'
 import generateModules from '../util/generateModules'
+import processPlugins from '../util/processPlugins'
 import wrapWithVariants from '../util/wrapWithVariants'
-
-function defineRule(selector, properties) {
-  const decls = _.map(properties, (value, property) => {
-    return postcss.decl({
-      prop: `${property}`,
-      value: `${value}`,
-    })
-  })
-
-  return postcss.rule({ selector }).append(decls)
-}
-
-function defineAtRule(atRule, rules) {
-  const [name, ...params] = atRule.split(' ')
-
-  return postcss
-    .atRule({
-      name: name.startsWith('@') ? name.slice(1) : name,
-      params: params.join(' '),
-    })
-    .append(rules)
-}
-
-function processPlugins(config) {
-  const pluginComponents = []
-  const pluginUtilities = []
-
-  config.plugins.forEach(plugin => {
-    plugin({
-      config,
-      rule: defineRule,
-      atRule: defineAtRule,
-      e: escapeClassName,
-      addUtilities: (utilities, variants) => {
-        pluginUtilities.push(wrapWithVariants(utilities, variants))
-      },
-      addComponents: components => {
-        pluginComponents.push(...components)
-      },
-    })
-  })
-
-  return [pluginComponents, pluginUtilities]
-}
 
 export default function(config) {
   return function(css) {
